remove separate threshold field, fixing validation when removing an owner wip
