+++ package.json
@@ -32,7 +32,7 @@
   "dependencies": {
     "@gnosis.pm/safe-contracts": "^1.0.0",
     "@gnosis.pm/util-contracts": "2.0.1",
-    "@material-ui/core": "4.2.0",
+    "@material-ui/core": "4.2.1",
     "@material-ui/icons": "4.2.1",
     "@testing-library/jest-dom": "^4.0.0",
     "@welldone-software/why-did-you-render": "3.2.1",
@@ -51,7 +51,7 @@
     "react-dom": "^16.8.6",
     "react-final-form": "6.3.0",
     "react-final-form-listeners": "^1.0.2",
-    "react-hot-loader": "4.12.7",
+    "react-hot-loader": "4.12.8",
     "react-infinite-scroll-component": "^4.5.2",
     "react-redux": "7.1.0",
     "react-router-dom": "^4.3.1",
@@ -63,9 +63,9 @@
     "web3": "1.0.0-beta.37"
   },
   "devDependencies": {
-    "@babel/cli": "7.5.0",
-    "@babel/core": "7.5.4",
-    "@babel/plugin-proposal-class-properties": "7.5.0",
+    "@babel/cli": "7.5.5",
+    "@babel/core": "7.5.5",
+    "@babel/plugin-proposal-class-properties": "7.5.5",
     "@babel/plugin-proposal-decorators": "7.4.4",
     "@babel/plugin-proposal-do-expressions": "7.5.0",
     "@babel/plugin-proposal-export-default-from": "7.5.2",
@@ -84,7 +84,7 @@
     "@babel/plugin-transform-member-expression-literals": "^7.2.0",
     "@babel/plugin-transform-property-literals": "^7.2.0",
     "@babel/polyfill": "7.4.4",
-    "@babel/preset-env": "7.5.4",
+    "@babel/preset-env": "7.5.5",
     "@babel/preset-flow": "^7.0.0-beta.40",
     "@babel/preset-react": "^7.0.0-beta.40",
     "@sambego/storybook-state": "^1.0.7",
@@ -102,18 +102,18 @@
     "babel-plugin-transform-es3-member-expression-literals": "^6.22.0",
     "babel-plugin-transform-es3-property-literals": "^6.22.0",
     "classnames": "^2.2.5",
-    "css-loader": "3.0.0",
+    "css-loader": "3.1.0",
     "detect-port": "^1.2.2",
     "eslint": "6.0.1",
     "eslint-config-airbnb": "17.1.1",
     "eslint-plugin-flowtype": "3.11.1",
     "eslint-plugin-import": "2.18.0",
-    "eslint-plugin-jest": "22.9.0",
+    "eslint-plugin-jest": "22.10.0",
     "eslint-plugin-jsx-a11y": "6.2.3",
     "eslint-plugin-react": "7.14.2",
     "ethereumjs-abi": "^0.6.7",
     "extract-text-webpack-plugin": "^4.0.0-beta.0",
-    "file-loader": "4.0.0",
+    "file-loader": "4.1.0",
     "flow-bin": "0.102.0",
     "fs-extra": "8.1.0",
     "html-loader": "^0.5.5",
@@ -123,7 +123,7 @@
     "json-loader": "^0.5.7",
     "mini-css-extract-plugin": "0.8.0",
     "postcss-loader": "^3.0.0",
-    "postcss-mixins": "^6.2.0",
+    "postcss-mixins": "6.2.2",
     "postcss-simple-vars": "^5.0.2",
     "pre-commit": "^1.2.2",
     "prettier-eslint-cli": "5.0.0",
@@ -135,7 +135,7 @@
     "truffle-contract": "4.0.24",
     "truffle-solidity-loader": "0.1.26",
     "uglifyjs-webpack-plugin": "2.1.3",
-    "webpack": "4.35.3",
+    "webpack": "4.36.1",
     "webpack-bundle-analyzer": "3.3.2",
     "webpack-cli": "3.3.6",
     "webpack-dev-server": "3.7.2",
